Make an AudioEncoder subclass for iLBC
BUG=3926
R=henrik.lundin@webrtc.org, kjellander@google.com
TBR=kjellander@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/32649005

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7828 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/modules/audio_coding/codecs/ilbc/interface/ilbc.h
@@ -138,7 +138,7 @@ extern "C" {
   int16_t WebRtcIlbcfix_Encode(iLBC_encinst_t *iLBCenc_inst,
                                const int16_t *speechIn,
                                int16_t len,
-                               int16_t *encoded);
+                               uint8_t* encoded);
 
   /****************************************************************************
    * WebRtcIlbcfix_DecoderInit(...)
